requirements.txt
# AWS Assistant Dependencies

# Core LangChain dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# OpenAI
openai>=1.0.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
colorama>=0.4.6

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0 